Renderer update
Added RenderRect with the ability to draw a rect with a fill or just a line, also set up the ground work for preventing the renderer from loading the same asset multiple times. Will likely need to find a good way to reference count the assets in that case, that can come later.